fix: add step functions in edit reservation page
